requirements.txt
langchain==0.2.5
langchain-community==0.2.5
deeplake==3.9.10
boto3==1.34.129
openai==1.34.0
cohere==5.5.8
tiktoken==0.7.0
beautifulsoup4==4.12.3
apify-client==1.7.0
python-dotenv==1.0.1
black==24.4.2
elevenlabs==1.3.1
streamlit==1.35.0
streamlit-chat==0.0.2.2
audio-recorder-streamlit==0.0.8